сломались db transactions и transactions

+++ 3. Асинхронность и веб/bank/backend/migrations/0001_init.sql
@@ -11,7 +11,7 @@ CREATE INDEX IF NOT EXISTS idx_users_email ON users (email);
 CREATE TABLE IF NOT EXISTS accounts (
     id UUID PRIMARY KEY,
     user_id UUID NOT NULL,
-    balance NUMERIC NOT NULL,
+    balance DOUBLE PRECISION NOT NULL,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(),
 
     CONSTRAINT fk_accounts_user
